pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mealplanner"
version = "0.1.0"
description = "A rule-based meal planning system"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "streamlit>=1.28",
    "pyyaml>=6.0",
    "fpdf2>=2.7",
]

[project.scripts]
mealplanner = "main:app"

[tool.hatch.build.targets.wheel]
packages = ["domain", "app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
